requirements.txt
python-dotenv
worldnewsapi
requests
flask
